OnlineBankingCZ: remove launchPdf logic from view into delegate
