+++ online-banking-cz/src/main/java/com/adyen/checkout/onlinebankingcz/DefaultOnlineBankingCZDelegate.kt
@@ -8,6 +8,8 @@
 
 package com.adyen.checkout.onlinebankingcz
 
+import android.content.Context
+import android.net.Uri
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.flow.MutableSingleEventSharedFlow
 import com.adyen.checkout.components.model.paymentmethods.PaymentMethod
@@ -15,6 +17,9 @@ import com.adyen.checkout.components.model.payments.request.OnlineBankingCZPayme
 import com.adyen.checkout.components.model.payments.request.PaymentComponentData
 import com.adyen.checkout.components.util.PaymentMethodTypes
 import com.adyen.checkout.core.exception.CheckoutException
+import com.adyen.checkout.core.exception.ComponentException
+import com.adyen.checkout.core.log.LogUtil
+import com.adyen.checkout.core.log.Logger
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -62,13 +67,22 @@ class DefaultOnlineBankingCZDelegate(
         _componentStateFlow.tryEmit(state)
     }
 
-    override fun getTermsAndConditionsUrl(): String = TERMS_CONDITIONS_URL
-
     fun onExceptionHappen(e: CheckoutException) {
         _exceptionFlow.tryEmit(e)
     }
 
     companion object {
+        private val TAG = LogUtil.getTag()
         private const val TERMS_CONDITIONS_URL = "https://static.payu.com/sites/terms/files/payu_privacy_policy_cs.pdf"
+    }
+
+    @Suppress("ReturnCount")
+    override fun launchOpenPdf(context: Context) {
+        val uri = Uri.parse(TERMS_CONDITIONS_URL)
+        if (OpenPdfUtils.launchNative(context, uri)) return
+        if (OpenPdfUtils.launchWithCustomTabs(context, uri)) return
+        if (OpenPdfUtils.launchBrowser(context, uri)) return
+        Logger.e(TAG, "Could not launch url")
+        throw ComponentException("failed to open terms and conditions pdf.")
     }
 }
