+++ online-banking-cz/src/main/java/com/adyen/checkout/onlinebankingcz/OnlineBankingCZView.kt
@@ -8,27 +8,16 @@
 
 package com.adyen.checkout.onlinebankingcz
 
-import android.content.ActivityNotFoundException
 import android.content.Context
-import android.content.Intent
-import android.content.Intent.ACTION_VIEW
-import android.net.Uri
-import android.os.Build
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.AdapterView
-import androidx.annotation.RequiresApi
-import androidx.browser.customtabs.CustomTabColorSchemeParams
-import androidx.browser.customtabs.CustomTabsIntent
 import androidx.lifecycle.LifecycleOwner
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.model.payments.request.OnlineBankingCZPaymentMethod
 import com.adyen.checkout.components.ui.adapter.SimpleTextListAdapter
-import com.adyen.checkout.components.ui.util.ThemeUtil
 import com.adyen.checkout.components.ui.view.AdyenLinearLayout
-import com.adyen.checkout.core.exception.CheckoutException
-import com.adyen.checkout.core.exception.ComponentException
 import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
 import com.adyen.checkout.onlinebankingcz.databinding.OnlineBankingCzSpinnerLayoutBinding
@@ -70,13 +59,7 @@ class OnlineBankingCZView @JvmOverloads constructor(
                 component.notifyInputDataChanged()
             }
         }
-        binding.textviewTermsAndConditions.setOnClickListener {
-            try {
-                launchOpenPdf()
-            } catch (e: CheckoutException) {
-                component.onExceptionHappen(e)
-            }
-        }
+        binding.textviewTermsAndConditions.setOnClickListener { component.openTermsAndConditionsPdf(context) }
     }
 
     override val isConfirmationRequired: Boolean
@@ -106,77 +89,6 @@ class OnlineBankingCZView @JvmOverloads constructor(
         super.setEnabled(enabled)
         binding.autoCompleteTextViewIssuers.isEnabled = enabled
         binding.textInputLayoutIssuers.isEnabled = enabled
-    }
-
-    private fun launchOpenPdf() {
-        val url = component.getTermsAndConditionsUrl()
-        if (url.isEmpty()) throw ComponentException("Terms and conditions file URL is empty.")
-        val uri = Uri.parse(url)
-        if (launchNative(uri)) return
-        if (launchWithCustomTabs(uri)) return
-        if (launchBrowser(uri)) return
-        Logger.e(TAG, "Could not launch url")
-        throw ComponentException("failed to open terms and conditions pdf.")
-    }
-
-    private fun launchNative(uri: Uri): Boolean {
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) launchNativeApi30(uri)
-        // because custom tabs pdf viewer is working before api 30 on chrome
-        else launchWithCustomTabs(uri)
-    }
-
-    @RequiresApi(Build.VERSION_CODES.R)
-    private fun launchNativeApi30(uri: Uri): Boolean {
-        val nativeAppIntent = Intent().apply {
-            action = ACTION_VIEW
-            setDataAndType(uri, "application/pdf")
-        }
-        return try {
-            context.startActivity(nativeAppIntent)
-            Logger.d(TAG, "launchNativeApi30 - open terms and conditions pdf successful with native app")
-            true
-        } catch (ex: ActivityNotFoundException) {
-            Logger.d(TAG, "launchNativeApi30 - could not find native app to terms and conditions pdf with", ex)
-            false
-        }
-    }
-
-    private fun launchWithCustomTabs(uri: Uri): Boolean {
-        // open in custom tabs if there's no native app for the target uri
-        val defaultColors = CustomTabColorSchemeParams.Builder()
-            .setToolbarColor(ThemeUtil.getPrimaryThemeColor(context))
-            .build()
-        return try {
-            CustomTabsIntent.Builder()
-                .setShowTitle(true)
-                .setDefaultColorSchemeParams(defaultColors)
-                .build()
-                .launchUrl(context, uri)
-            Logger.d(TAG, "launchWithCustomTabs - open terms and conditions pdf successful with custom tabs")
-            true
-        } catch (e: ActivityNotFoundException) {
-            Logger.d(TAG, "launchWithCustomTabs - device doesn't support custom tabs or chrome is disabled", e)
-            false
-        }
-    }
-
-    /**
-     * in case the device doesn't support custom tabs or doesn't support google services (Huawei device).
-     */
-    private fun launchBrowser(uri: Uri): Boolean {
-        return try {
-            val browserActivityIntent = Intent()
-                .setAction(ACTION_VIEW)
-                .addCategory(Intent.CATEGORY_BROWSABLE)
-                .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                .setData(uri)
-            context.startActivity(browserActivityIntent)
-            Logger.d(TAG, "launchBrowser - open terms and conditions pdf successful with browser")
-            true
-        } catch (e: ActivityNotFoundException) {
-            Logger.d(TAG, "launchBrowser - could not open pdf on browser or there's no browser!", e)
-            false
-        }
     }
 
     companion object {
